+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/debug/renderer/OptionsRendererTest.groovy
@@ -1,0 +1,160 @@
+package ru.vyarus.dropwizard.guice.config.debug.renderer
+
+import io.dropwizard.Application
+import io.dropwizard.Configuration
+import io.dropwizard.setup.Bootstrap
+import io.dropwizard.setup.Environment
+import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.module.context.debug.DiagnosticBundle
+import ru.vyarus.dropwizard.guice.module.context.debug.report.option.OptionsRenderer
+import ru.vyarus.dropwizard.guice.module.context.option.Option
+import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBootstrap
+import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBundle
+import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
+import spock.lang.Specification
+
+import javax.inject.Inject
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 13.08.2016
+ */
+@UseGuiceyApp(App)
+class OptionsRendererTest extends Specification {
+
+    @Inject
+    OptionsRenderer renderer
+
+    def "Check all render"() {
+
+        expect:
+        render(true) == """
+
+    BundleX                   (r.v.d.g.c.d.r.OptionsRendererTest\$BundleXOptions)
+        DebugFeature                   = false
+        StrictMode                     = true                           *CUSTOM
+        NullValue                      = null
+        ListValue                      = [one, two]
+
+
+    Guicey                    (r.v.dropwizard.guice.GuiceyOptions)
+        ScanPackages                   = [com.foo, com.bat]             *CUSTOM
+        SearchCommands                 = false
+        UseCoreInstallers              = true
+        ConfigureFromDropwizardBundles = false
+        InjectorStage                  = PRODUCTION
+
+
+    OtherOpts                 (r.v.d.g.c.d.r.OptionsRendererTest\$OtherOpts)
+        Opt1                           = sample                         *CUSTOM, NOT_USED
+""" as String;
+    }
+
+
+    def "Check not used marker disable"() {
+
+        expect:
+        render(false) == """
+
+    BundleX                   (r.v.d.g.c.d.r.OptionsRendererTest\$BundleXOptions)
+        DebugFeature                   = false
+        StrictMode                     = true                           *CUSTOM
+        NullValue                      = null
+        ListValue                      = [one, two]
+
+
+    Guicey                    (r.v.dropwizard.guice.GuiceyOptions)
+        ScanPackages                   = [com.foo, com.bat]             *CUSTOM
+        SearchCommands                 = false
+        UseCoreInstallers              = true
+        ConfigureFromDropwizardBundles = false
+        InjectorStage                  = PRODUCTION
+
+
+    OtherOpts                 (r.v.d.g.c.d.r.OptionsRendererTest\$OtherOpts)
+        Opt1                           = sample                         *CUSTOM
+""" as String;
+    }
+
+    String render(boolean config) {
+        renderer.renderReport(config).replaceAll("\r", "").replaceAll(" +\n", "\n")
+    }
+
+    static class App extends Application<Configuration> {
+
+        @Override
+        void initialize(Bootstrap<Configuration> bootstrap) {
+            bootstrap.addBundle(GuiceBundle.builder()
+                    .enableAutoConfig("com.foo", "com.bat")
+                    // never used value
+                    .option(OtherOpts.Opt1, "sample")
+                    .option(BundleXOptions.StrictMode, true)
+                    .modules(new DiagnosticBundle.DiagnosticModule())
+                    .bundles(new GuiceyBundle() {
+                @Override
+                void initialize(GuiceyBootstrap bs) {
+                    bs.option(BundleXOptions.DebugFeature)
+                    bs.option(BundleXOptions.StrictMode)
+                    bs.option(BundleXOptions.NullValue)
+                    bs.option(BundleXOptions.ListValue)
+                }
+            })
+                    .build())
+        }
+
+        @Override
+        void run(Configuration configuration, Environment environment) throws Exception {
+
+        }
+    }
+
+    static enum OtherOpts implements Option {
+
+        Opt1(String, "foo"),
+        Opt2(String, "bar");
+
+        Class type
+        Object value
+
+        OtherOpts(Class type, Object value) {
+            this.type = type
+            this.value = value
+        }
+
+        @Override
+        Class getType() {
+            return type
+        }
+
+        @Override
+        Object getDefaultValue() {
+            return value
+        }
+    }
+
+    static enum BundleXOptions implements Option {
+
+        DebugFeature(Boolean, false),
+        StrictMode(Boolean, true),
+        NullValue(String, null),
+        ListValue(List.class, Arrays.asList("one", "two"));
+
+        Class type
+        Object value
+
+        BundleXOptions(Class type, Object value) {
+            this.type = type
+            this.value = value
+        }
+
+        @Override
+        Class getType() {
+            return type
+        }
+
+        @Override
+        Object getDefaultValue() {
+            return value
+        }
+    }
+}